updated dependencies; moved most types to types.ts; moved locale type workaround to dedicated file
