+++ package.json
@@ -60,21 +60,23 @@
   "author": "Hauke Broer <info@owja.de>",
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^25.2.3",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
+    "@types/jest": "^26.0.10",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
     "cross-env": "^7.0.2",
-    "eslint": "^6.8.0",
+    "eslint": "^7.7.0",
     "eslint-plugin-jest": "^23.20.0",
     "full-icu": "^1.3.1",
-    "iso-639-1": "^2.1.3",
-    "jest": "^25.5.4",
+    "iso-639-1": "^2.1.4",
+    "jest": "^26.4.1",
     "microbundle": "^0.12.3",
     "prettier": "^2.0.5",
     "shx": "^0.3.2",
-    "ts-jest": "^25.5.1",
+    "ts-jest": "^26.2.0",
     "ts-node": "^8.10.2",
-    "typescript": "^3.9.7",
+    "typescript": "^4.0.2"
+  },
+  "optionalDependencies": {
     "@formatjs/intl-locale": "^2.3.0"
   },
   "mangle": {
