+++ src/lib/translator.ts
@@ -1,54 +1,15 @@
 import {_maximize, _parse, _validateLanguageTag} from "./utils";
-
-export type TranslateFunction = (key: string, options?: Partial<TranslateOptions>) => string;
-
-export interface TranslatorInterface {
-    t: TranslateFunction;
-    /** @deprecated use locale() instead */
-    language(language?: string): string;
-    locale(locale?: string | Intl.Locale): void;
-    short(): string;
-    long(): string;
-    region(): string;
-    script(): string | undefined;
-    addResource(language: string, translations: Translations): void;
-    addPlugin(plugin: TranslatorPlugin, language?: string): void;
-    listen(listener: Listener): Unsubscribe;
-}
-
-export type Translations = {
-    [key: string]: string | Translations;
-};
-
-export type ParsedTranslations = {
-    [search: string]: string;
-};
-
-export type TranslateOptions = {
-    context: string;
-    count: number;
-    replace: {
-        [search: string]: string | number | Date;
-    };
-};
-
-export type LanguageOptions = {
-    default: string;
-    fallback: string;
-};
-
-export type Listener = () => void;
-export type Unsubscribe = () => void;
-
-export type TranslatorPlugin = (
-    translated: string,
-    options: Partial<TranslateOptions>,
-    translator: TranslatorInterface,
-) => string | undefined;
-
-export type PluginRegistry = {
-    [tag: string]: TranslatorPlugin[];
-};
+import {
+    LanguageOptions,
+    Listener,
+    ParsedTranslations,
+    PluginRegistry,
+    TranslateOptions,
+    Translations,
+    TranslatorInterface,
+    TranslatorPlugin,
+    Unsubscribe,
+} from "./types";
 
 const GLOBAL = "global";
 
@@ -176,7 +137,7 @@ export class Translator implements TranslatorInterface {
     /**
      * Add translations for a language tag
      */
-    addResource(languageTag: string, translations: Translations) {
+    addResource(languageTag: string, translations: Translations): void {
         languageTag = _validateLanguageTag(languageTag);
         this._resources = {...this._resources, ..._parse(translations, languageTag)};
         this._trigger();
@@ -185,7 +146,7 @@ export class Translator implements TranslatorInterface {
     /**
      * Add a plugin for a language tag
      */
-    addPlugin(plugin: TranslatorPlugin, languageTag = GLOBAL) {
+    addPlugin(plugin: TranslatorPlugin, languageTag = GLOBAL): void {
         if (languageTag !== GLOBAL) {
             languageTag = _validateLanguageTag(languageTag);
         }
