+++ src/plugins.ts
@@ -1,3 +1,5 @@
+import "./lib/intl.locale";
+
 export {createCurrencyPlugin} from "./lib/currency-plugin";
 export {createDateTimePlugin} from "./lib/date-time-plugin";
